+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Moved requirement to have at least 2 CPU cores to recommendations, as 4Gb RAM now.
 - `Response.HeadersDictionary` is gone and replaced with function `Response.SetHeadersDictionary`. To set individual headers use `response["HeaderName"] = "HeaderValue"` syntax.
 - `HandlerOptions.ProxyDelegateTrigger` is now an internal flag that is no longer exposed to user applications (affected Launcher app).
+- Starcounter is always installed in a subfolder called "Starcounter".
+- Version number in the installation path was removed.
 
 ### Fixed
 - Unhandled exceptions in UDP/TCP handlers:
@@ -26,3 +28,4 @@ https://github.com/Starcounter/Starcounter/issues/2886
 - Bug fixed regarding headers dictionary creation (CreateHeadersDictionaryFromHeadersString):
 [#2939](https://github.com/Starcounter/Starcounter/issues/2939).
 - Fixed extraction of CRT libraries in installer GUI that caused the issue [#2759](https://github.com/Starcounter/Starcounter/issues/2759)
+- Bug fixed when handling error from indexcreation, that caused an assertion failure instead of returning the error to usercode: [#2951](https://github.com/Starcounter/Starcounter/issues/2951)